postsocial/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery==5.4.0
redis==5.2.1
flower==2.0.1
httpx==0.28.1
pillow==11.1.0
cryptography==44.0.0
email-validator==2.2.0
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
